.history/backend/requirements_20220630073350.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
idna==3.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
